% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013_sample}
\alias{accident_2013_sample}
\title{Sample of 2013 FARS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 50 columns.
}
\usage{
accident_2013_sample
}
\description{
A dataset containing the first 20 rows of the FARS data
from 2013. Used for vignette example.
}
\keyword{datasets}
