% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_function.R
\name{browse_color_discrete}
\alias{browse_color_discrete}
\title{Browse Discrete color}
\usage{
browse_color_discrete(mycolor)
}
\arguments{
\item{mycolor}{Your custom colors(a vector).}
}
\value{
A ggplot2 object.
}
\description{
Use a bar chart to display your discrete colors.
}
\examples{
browse_color_discrete(mycolor)
\donttest{
browse_color_discrete(mycolor)}
}
