% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_function.R
\name{browse_color_gradient}
\alias{browse_color_gradient}
\title{Browse gradient color}
\usage{
browse_color_gradient(mycolor)
}
\arguments{
\item{mycolor}{Your custom colors(a vector)}
}
\value{
A ggplot2 object.
}
\description{
Browse gradient color
}
\examples{
browse_color_gradient(mycolor)
\donttest{
browse_color_gradient(mycolor)}
}
