% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate mean function}
\usage{
calc_mean(v)
}
\arguments{
\item{v}{vector}
}
\value{
mean
}
\description{
This function allows you to calculate the mean of a sequence vector. It interpolates the missing values.
}
\examples{
calc_mean(c(1,3,NA,6,7))

}
