% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-color-themes.R
\name{custom_palette}
\alias{custom_palette}
\title{Custom color palettes for \code{ggplot2}}
\usage{
custom_palette(
  palette = c("viridis", "high_contrast", "redblue"),
  viridis_option = "plasma"
)
}
\arguments{
\item{palette}{Character. The name of the color palette to use. Options are
"viridis", "high_contrast", and "redblue".}

\item{viridis_option}{Character. The viridis color palette option to use.
Only used if \code{palette == "viridis"}. Default is "plasma".
See \code{\link[viridis]{viridis}} for more information.}
}
\value{
A function that generates a color palette with the specified number
of colors.
}
\description{
Generates various customized color palettes for \code{ggplot2}
}
\examples{
library(scales)
show_col(custom_palette(palette = "high_contrast")(8))
show_col(custom_palette(palette = "redblue")(5))
show_col(custom_palette(palette = "viridis")(8))

}
