% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_umap_and_tsne.R
\name{d_tsne}
\alias{d_tsne}
\title{Enhance UMAP}
\usage{
d_tsne(obj = NULL)
}
\arguments{
\item{obj}{obj Seurat object}
}
\value{
A new style of plot with arrows
}
\description{
Enhance UMAP
}
\examples{
+ d_tsne(obj)
\donttest{
+ d_tsne(obj)}

}
\author{
dk
}
