% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_umap_and_tsne.R
\name{d_umap}
\alias{d_umap}
\title{Enhance UMAP}
\usage{
d_umap(obj = NULL)
}
\arguments{
\item{obj}{Seurat object \code{\link[base]{sum}}}
}
\value{
A new style of plot with arrows
}
\description{
Further beautify plot based on UMAP
this is a test function
}
\examples{
+ d_umap(obj)
\donttest{
+ d_umap(obj)}

}
