% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Data Input}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{.csv file from which the data will be read}
}
\value{
heck if the file exists, gives error if it does not. if the file exists,
 it read the file in a data frame and returns it \code{getwd()}.
}
\description{
After checking if a file exist in a given path, the function read the
 file and creat a data frame table.
}
\examples{
\dontrun{
  fars_read("accident_2010.csv.bz2")
}

}
