% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthly number of accident for given years in FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years for which FARS data file will be loaded and number
of accident will be summurized \code{years}}
}
\value{
return NULL for an invalid year. For a valid input, returns list of
FARS data in data frame in years and number of accidents. It uses \code{\link{fars_read_years}} function.
}
\description{
Monthly number of accident for given years in FARS data
}
\examples{
  \dontrun{
    fars_summarize_years(c(2013, 2014))
    fars_summarize_years(c("2013", "2014"))
  }
}
