% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{fibonacci}
\alias{fibonacci}
\title{Return the nth element in the Fibonacci sequence}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{a non-zero integer}
}
\value{
nth Fibonacci element in sequence
}
\description{
Return the nth element in the Fibonacci sequence
}
