% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_lm_analysis.R
\name{hw_lm_analysis}
\alias{hw_lm_analysis}
\title{Height vs Weight Linear Regression Analysis}
\usage{
hw_lm_analysis(df, plot = TRUE)
}
\arguments{
\item{df}{A data frame with columns: height and weight}

\item{plot}{Logical; if TRUE, plot scatter plot with regression line}
}
\value{
A list with p-value and conclusion
}
\description{
Height vs Weight Linear Regression Analysis
}
