% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISC}
\alias{ISC}
\title{RPKM value of ISC cells from fly five gut region RNAseq}
\format{
data frame, with 15683 rows and 5 variables:
\describe{
\item{R1}{Gut Region 1: numeric}
\item{R2}{Gut Region 2: numeric}
\item{R3}{Gut Region 3: numeric}
\item{R4}{Gut Region 4: numeric}
\item{R5}{Gut Region 5: numeric}
}
}
\source{
\url{http://flygutseq.buchonlab.com/resources}
}
\usage{
ISC
}
\description{
A dataset containing the RPKM value from fly five gut region RNAseq.
}
\keyword{datasets}
