% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creat a file name from the data source for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Only logical, integer, real and character vectors are supported. The
year for which the file will be created}
}
\value{
read pram as integer. and return file name for a given year
}
\description{
Creat a file name from the data source for a given year
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2013")
}
}
