% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_logistic_regression.R
\name{manual_logistic_regression}
\alias{manual_logistic_regression}
\title{manual_logistic_regression}
\usage{
manual_logistic_regression(x, y, threshold = 1e-10, max_iter = 100)
}
\arguments{
\item{x}{input value}
}
\value{
the square of x
}
\description{
This is an function named 'manual_logistic_regression' which xxx
}
\examples{
square(3)

}
