% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753FALLfaug.R
\docType{package}
\name{MATH4753FALLfaug}
\alias{MATH4753FALLfaug-package}
\alias{MATH4753FALLfaug}
\title{Lab Functions Package}
\description{
Contains the functions from MATH4753 Labs
}
\author{
Noah Faught
}
