% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_na.R
\name{mean_na}
\alias{mean_na}
\title{Calculate Mean Ignoring NA Values}
\usage{
mean_na(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the numeric vector excluding NA values, or NA if the vector contains only NA values.
}
\description{
This function calculates the mean of a numeric vector, ignoring any NA values. If the entire vector is NA, it returns NA.
}
\examples{
mean_na(c(1, 2, 3, NA, 4))
mean_na(c(NA, NA, NA, NA, NA))
}
