% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Iteration Barplot}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{"description"}

\item{n}{"description"}

\item{p}{"description"}
}
\value{
A sophisticated barplot
}
\description{
Iteration Barplot
}
\examples{
\dontrun{bin100 = mybin(iter = 100, n = 10, p = 0.7)}
}
