% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of bootstrap resembles}

\item{x}{numeric vector}

\item{fun}{statistical function}

\item{alpha}{significance level}

\item{cx}{scaling factor}

\item{...}{additional arguments}
}
\value{
a sophisticated graph
}
\description{
myboot2
}
\examples{
obj<- MATH4753FALLfaug::myboot2(x=fire$DISTANCE, fun = "mean", xlab = "mean")
}
