% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
confidence interval
}
\description{
myci
}
\examples{
set.seed(23);x = rnorm(30,mean=10,sd=12)
MATH4753FALLfaug::myci(x)

}
