% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{rate parameter for poisson distribution}

\item{...}{additional arguments}
}
\value{
a sophisticated graph
}
\description{
mycltp
}
\examples{
mycltp(n = 5, iter = 10000)

}
