% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{mycurve}
\alias{mycurve}
\title{Curved Graph}
\usage{
mycurve(x, m, l)
}
\arguments{
\item{x}{x}

\item{m}{mean}

\item{l}{standard deviation}
}
\value{
A sophisticated curved plot
}
\description{
Curved Graph
}
\examples{
mycurve(5,10,3)
}
