% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{myLasso}
\alias{myLasso}
\title{Find the lasso solution path for various values of 
the regularization parameter lambda.}
\usage{
myLasso(X, Y, lambda_all)
}
\arguments{
\item{X:}{n x p matrix of explanatory variables.}

\item{Y:}{n dimensional response vector}

\item{lambda_all:}{Vector of regularization parameters.}
}
\value{
Returns a matrix containing the lasso solution vector 
beta of size (p+1) x length(lambda_all) for each 
regularization parameter.
}
\description{
Find the lasso solution path for various values of 
the regularization parameter lambda.
}
\examples{
myLasso(X, Y, lambda_all)
}
