% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{myLM}
\alias{myLM}
\title{Perform the linear regression of Y on X}
\usage{
myLM(X, Y, pFlag = 0)
}
\arguments{
\item{X}{is an n x p matrix of explanatory variables}

\item{Y}{is an n dimensional vector of responses}

\item{pFlag}{is whether to consider intercept (default 0)}
}
\value{
least squares solution vector beta "beta" and standard error vector "std"
}
\description{
Perform the linear regression of Y on X
}
\examples{
myLM(X,Y,1)
}
