% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{myLogistic}
\alias{myLogistic}
\title{Perform logistic regression of Y on X}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{is an n x p matrix of explanatory variables}

\item{Y}{is an n dimensional vector of responses}
}
\value{
logistic regression solution vector beta "beta" and standard error vector "std"
}
\description{
Perform logistic regression of Y on X
}
\examples{
myLogistic(X,Y)
}
