% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample vector}

\item{mu}{values for the mean}

\item{sig}{candidate values for the standard deviation}

\item{...}{additional arguments}
}
\value{
a sophisticated graph
}
\description{
mymlnorm
}
\examples{
mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
