% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{myQR}
\alias{myQR}
\title{QR decomposition}
\usage{
myQR(A)
}
\arguments{
\item{A, }{an n x m matrix}
}
\value{
a list with Q.transpose and R, Q is an orthogonal n x n matrix,
R is an upper triangular n x m matrix, A = Q %*% R
}
\description{
QR decomposition
}
\examples{
myQR(A)
}
