% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{myRidge}
\alias{myRidge}
\title{Perform ridge regression of Y on X using lambda}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{X}{is an n x p matrix of explanatory variables}

\item{Y}{is an n dimensional vector of dependent variables}

\item{lambda:}{regularization parameter (lambda >= 0)}
}
\value{
beta, the (p+1) ridge regression solution
}
\description{
Perform ridge regression of Y on X using lambda
}
\examples{
myRidge(X,Y,lambda)
}
