\name{myrpackage-package}
\alias{myrpackage-package}
\alias{myrpackage}
\docType{package}
\title{
  A Multilingual Greeting Package
}
\description{
  A friendly R package that provides greeting and farewell functions in multiple languages.
  This package demonstrates proper R package structure, documentation, and test practices.
  It is designed to be a simple but complete example of R package development.
}
\section{Main functions}{
  \itemize{
    \item \code{\link{hello}} - Say hello in various languages
    \item \code{\link{goodbye}} - Say goodbye in various languages
  }
}
\section{Supported languages}{
  \itemize{
    \item English (default)
    \item Spanish
    \item French
    \item Portuguese
    \item German
    \item Italian
  }
}
\author{
  Diogo Ribeiro <dfr@esmad.ipp.pt>
}
\keyword{package}
