% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{mySpline}
\alias{mySpline}
\title{Perform spline regression of Y on X using lambda and p.}
\usage{
mySpline(x, Y, lambda, p = 100)
}
\arguments{
\item{x:}{An n x 1 vector or matrix of explanatory variables.}

\item{Y:}{An n x 1 vector of dependent variables. Y can also be a 
matrix, as long as the function works.}

\item{lambda:}{regularization parameter (lambda >= 0)}

\item{p:}{Number of cuts to make to the x-axis.}
}
\value{
a list containing two elements:
The first element of the list is the spline regression
beta vector, which should be p + 1 dimensional.
The second element is y.hat, the predicted Y values
using the spline regression beta vector. This 
can be a numeric vector or matrix.
}
\description{
Perform spline regression of Y on X using lambda and p.
}
\examples{
mySpline(x,Y,lambda,p)
}
