% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRFunctions.R
\name{mySweep}
\alias{mySweep}
\title{Perform a SWEEP operation on A with the pivot element A[m,m].}
\usage{
mySweep(A, m)
}
\arguments{
\item{A, }{an square matrix; m, the pivot element index.}
}
\value{
a swept matrix
}
\description{
Perform a SWEEP operation on A with the pivot element A[m,m].
}
\examples{
mySweep(A,m)
}
