% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{probability a plane will truly be overbooked}

\item{p}{probability of a "show"}
}
\value{
A list of values and sophisticated graphs
}
\description{
ntickets
}
\examples{
result <- ntickets(N = 200, gamma = 0.02, p = 0.95)
print(result)
}
