% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_printing_summary_plot.R
\name{plot.hdi_data}
\alias{plot.hdi_data}
\title{Plot method for hdi_data}
\usage{
\method{plot}{hdi_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hdi_data}.}

\item{...}{Additional arguments for the plot function.}
}
\description{
Plot method for hdi_data
}
