% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_printing_summary_plot.R
\name{print.hdi_data}
\alias{print.hdi_data}
\title{Print method for hdi_data}
\usage{
\method{print}{hdi_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hdi_data}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for hdi_data
}
