% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly Printer function for a tibble}
\usage{
printer(r, x)
}
\arguments{
\item{r}{what you want in the second column}

\item{x}{what you want in the first column}
}
\value{
a tibble
}
\description{
Silly Printer function for a tibble
}
\examples{
printer(x=rnorm(5), r = rnorm(5))
}
