% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdi.R
\name{read_hdi}
\alias{read_hdi}
\title{Read and Process Human Development Indicators Data}
\usage{
read_hdi(file)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file.}
}
\value{
A data.frame object.
}
\description{
This function reads the Human Development Indicators data from a CSV file,
processes the data by setting appropriate column types.
}
\examples{
file <- system.file("extdata", "hdro_indicators_chn.csv", package = "myRpackage")
hdi_data <- read_hdi(file)
}
