% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_data.R
\name{sample_read}
\alias{sample_read}
\title{Read CSV Duplicate}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{tibble}
}
\description{
Read CSV Duplicate
}
\examples{
csv=system.file("esri", "esri.csv", package="myrpackage")
sample_read(csv)
}
