% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA}
\alias{scRNA}
\title{A subset scRNAseq data in Seurat object}
\format{
Seurat object, 1869 cells and 17559 features.
}
\source{
\url{http://www.nibs.ac.cn/en/yjsjyimgshow.php?cid=5&sid=6&id=774}
}
\usage{
scRNA
}
\description{
Fly gut EEs scRNAseq.
}
\keyword{datasets}
