% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Compare}
\alias{scRNAseq_Score_Compare}
\title{Compare scRNAseq_Score_Region and scRNAseq_Score_Region2 outputs by correlation method}
\usage{
scRNAseq_Score_Compare(
  ScoreList,
  ScoreMatrix,
  Method = c("spearman", "pearson", "kendall")
)
}
\arguments{
\item{ScoreList}{A list: Output from scRNAseq_Score_Region function}

\item{ScoreMatrix}{A matrix: Output from scRNAseq_Score_Region2 function}

\item{Method}{method used in cor function, one of "spearman", "pearson", "kendall".}
}
\value{
A named Vector
}
\description{
Compare scRNAseq_Score_Region and scRNAseq_Score_Region2 outputs by correlation method
}
\examples{
data(scRNA)
data(bulkRNA)
score.list <- scRNAseq_Score_Region(scRNA, bulkRNA)
score.matrix <- scRNAseq_Score_Region2(scRNA, bulkRNA, Method = "spearman", Genes.Selection = "Top")
scRNAseq_Score_Compare(score.list,score.matrix)
}
