% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Region}
\alias{scRNAseq_Score_Region}
\title{Calculate Region preference for each cluster}
\usage{
scRNAseq_Score_Region(
  SeuratObj,
  BulkRNAseq.expr,
  UMI.gradient = c(10, 20, 30, 40, 50, 100, 200, 500, 1000, 1500, 2000),
  Genes.gradient = c(10, 20, 30, 40, 50, 100, 200, 300, 400, 500),
  minimal.RNAseq.value = 0.5,
  scRNAseq.expression.cut = 0.5
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{BulkRNAseq.expr}{A data frame: Bulk RNA-seq gene expression values of each region, should have same gene names(row names) as in scRNA-seq}

\item{UMI.gradient}{A numeric vector: UMI cut-off for selecting genes with enough summed expression}

\item{Genes.gradient}{A numeric vector: selected genes numbers for all regions.}

\item{minimal.RNAseq.value}{A numeric: remove genes with low expression value in bulk RNA-seq}

\item{scRNAseq.expression.cut}{A numeric: cut off of the scaled value to define whether a gene is expressed or not.}
}
\value{
A list: cluster * Region preference for each UMI And Top N genes combination
}
\description{
Calculate Region preference for each cluster
}
\examples{
data(scRNA)
data(bulkRNA)
score.list <- scRNAseq_Score_Region(scRNA, bulkRNA)
}
