% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize Numeric Vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A standardized numeric vector with mean zero and standard deviation one.
}
\description{
This function standardizes a numeric vector by subtracting the mean and dividing by the standard deviation. This transformation results in a vector with a mean of zero and a standard deviation of one.
}
\examples{
standardize(c(1, 2, 3, 4, 5, NA))
}
