% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_printing_summary_plot.R
\name{summary.hdi_data}
\alias{summary.hdi_data}
\title{Summary method for hdi_data}
\usage{
\method{summary}{hdi_data}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hdi_data}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary method for hdi_data
}
