% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{test}
\alias{test}
\title{A test function}
\usage{
test()
}
\value{
Just print some word!
}
\description{
A test function
}
\examples{
## test()
## nothing of just a test function, which will print some words!
}
