% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_na.R
\name{variance_na}
\alias{variance_na}
\title{Calculate Variance Ignoring NA Values}
\usage{
variance_na(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The variance of the numeric vector excluding NA values, or NA if the vector contains only NA values.
}
\description{
This function calculates the variance of a numeric vector, ignoring any NA values. If the entire vector is NA, it returns NA.
}
\examples{
variance_na(c(1, 2, 3, NA, 4))
variance_na(c(NA, NA, NA))
}
