% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulkRNA}
\alias{bulkRNA}
\title{RPKM value from fly five gut region RNAseq}
\format{
A data frame with 15287 rows and 5 variables:
\describe{
\item{R1}{Gut Region 1}
\item{R2}{Gut Region 2}
\item{R3}{Gut Region 3}
\item{R4}{Gut Region 4}
\item{R5}{Gut Region 5}
}
}
\source{
\url{http://flygutseq.buchonlab.com/resources}
}
\usage{
bulkRNA
}
\description{
A dataset containing the RPKM value from fly five gut region RNAseq.
}
\keyword{datasets}
