% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to the file}
}
\value{
a tibble
}
\description{
Reads a .csv file containing FARS data into a tibble
}
\details{
This function takes the path to a file containing FARS data. The FARS data
is in .csv.bz2 format, which will be decompressed and read automatically into a
tibble.

Requires the \code{readr} and \code{tibble} packages.
}
\section{Errors}{

Will raise an error if the file is not found
}

\examples{
\dontrun{
fars_data <- fars_read("path/to/fars_data.csv")
}
}
\seealso{
\code{\link[=fars_read_years]{fars_read_years()}}, \code{\link[=fars_summarize_years]{fars_summarize_years()}}, \code{\link[=fars_map_state]{fars_map_state()}}
}
