% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Pre-format dataframe for selection}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years}
}
\value{
a list of tribbles with two columns, MONTH and year
}
\description{
Pre-format dataframe for selection
}
\details{
This function takes a list of years and attempts to read the FARS
data file into a data frame for each year. The year is then appended
as a column to each data frame, Then it along with the MONTH column
are extracted. The result is a 2-column data frame with the same number
of rows as the original FARS data for that year containing the MONTH column
from the FARS data and the year of the data file.

This function must be called from the directory where the
data is located.

This function requires \code{dplyr}
}
\section{Errors}{

This function will raise an error if a FARS-formatted file
for that year cannot be located in the current directory
}

\examples{
\dontrun{
years <- fars_read_years(list(2013, 2014))
}
}
\seealso{
\code{\link[=fars_summarize_years]{fars_summarize_years()}}
}
