% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Returns monthly number of accidents by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{list of years}
}
\value{
data frame whose first column is the month and whose subsequent columns
contain monthy accident counts for each year passed to the function
}
\description{
Returns monthly number of accidents by year
}
\details{
This function takes in a list of years and returns a data frame
containing the monthly number of accidents for each year. The first column
of the data frame contains the month, with subsequent columns each corresponding to
a year and each row the number of accidents that occurred in that month for that year.

This function requires \code{dplyr} and \code{tidyr}.
}
\examples{
\dontrun{
annual_summary <- fars_summarize_years(list(2013, 2014))
}

}
\seealso{

}
