% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{YD Hello World Function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hello}
}
\value{
The output from \code{\link{print}}
}
\description{
YD Hello World Function
}
\examples{
hello ("Yudong")
\dontrun{
hello("CRC")
}
}
