% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year}
}
\value{
a valid FARS data file name
}
\description{
Generates the properly formatted file name for FARS dataset based on the year.
Filenames are in the form "accident_<year>.csv.bz2".
}
\examples{
\dontrun{
filename <- make_filename(2013)
}

}
\seealso{
\code{\link[=fars_read_years]{fars_read_years()}}, \code{\link[=fars_map_state]{fars_map_state()}}
}
