% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_logistic_regression.R
\name{manual_logistic_regression}
\alias{manual_logistic_regression}
\title{manual_logistic_regression}
\usage{
manual_logistic_regression(x, y, threshold = 1e-10, max_iter = 100)
}
\arguments{
\item{x}{input value, predictors preferred in a matrix form}

\item{y}{input value, response vector consists only of 0 and 1}

\item{threshold}{input value, a numeric value checking if the logistic regression algorithm is
converging or not (the amount of beta changes should get smaller and smaller with each iteration,
eventually falling below the threshold), default threshold = 1e-10}

\item{max_iter}{input value, the maximum number of iterations, default max_iter = 100}
}
\value{
A list containing model summary information, including:

coefficients: estimated parameters for intercept and each variable (predictor),
according to the response

covariance_matrix: used to derive the standard errors and confidence intervals of the
fitted model's coefficient estimates

Dispersion: dispersion parameter for binomial family taken to be 1

standard_error: standard error for intercept and each variable (predictor),
according to the response

t_statistics: test statistics for intercept and each variable (predictor),
according to the response

p_value: P-value for intercept and each variable (predictor), according to the response

deviance_residuals: measuring how much probabilities estimated from fitted model differ
from the observed proportions of successes

quantile_deviance_residuals: a five-number summary of the residuals, including Minimum,
1st Quantile, Median, 3rd Quantile and Maximum, respectively

Null_deviance_degrees_of_freedom: the degrees of freedom of Null deviance

Residual_deviance_degrees_of_freedom: the degrees of freedom of Residual deviance

Null_deviance: measuring how well the response variable can be predicted by a model with
only an intercept term

Residual_deviance: measuring how well the response variable is predicted by the model when the
predictors are included

AIC: Akaike Information Criterion

number_iterations: the number of iterations
}
\description{
This is an function named 'manual_logistic_regression' which implements a logistic regression
used for modelling categorical data. This function assuming the response variable Y
falls into exactly two categories, Y = 1 and the other by Y = 0.
}
\examples{
set.seed(2022)
x1 = rnorm(30,3,2) + 0.1*c(1:30)
x2 = rbinom(30, 1,0.3)
x3 = rpois(n = 30, lambda = 4)
x3[16:30] = x3[16:30] - rpois(n = 15, lambda = 2)
x = cbind(x1,x2,x3)
y = c(rbinom(5, 1,0.1),rbinom(10, 1,0.25),rbinom(10, 1,0.75),rbinom(5, 1,0.9))
manual_glm <- manual_logistic_regression(x, y)

}
